Moved to 'token' package
